Exec integr

+++ firmware/drivers/wifiComms/file_server.c
@@ -21,9 +21,10 @@
 
 #include "esp_err.h"
 #include "esp_log.h"
-
 #include "esp_vfs.h"
 #include "esp_http_server.h"
+
+#include "robot_data.h"
 
 /* Max length a file path can have on storage */
 #define FILE_PATH_MAX (ESP_VFS_PATH_MAX + CONFIG_SPIFFS_OBJ_NAME_LEN)
@@ -37,25 +38,6 @@
 #define SCRATCH_BUFSIZE  8192
 
 #define DATA_BUFSIZE 128
-typedef enum {
-    STOP,
-    START,
-    PAUSE
-}robot_state;
-
-typedef struct {
-    uint8_t base_angle;
-    uint8_t arm_angle;
-    uint8_t tool_angle;
-    uint8_t x_coord;
-    uint8_t y_coord;
-} controls;
-
-typedef struct {
-    uint16_t x_accel;
-    uint16_t y_accel;
-    uint16_t z_accel;
-} accel_data;
 
 struct file_server_data {
     /* Base path of file storage */
@@ -71,7 +53,7 @@ static const char *TAG = "file_server";
  * a list of all files and folders under the requested path.
  * In case of SPIFFS this returns empty list when path is any
  * string other than '/', since SPIFFS doesn't support directories */
-static esp_err_t http_resp_dir_html(httpd_req_t *req, const char *dirpath)
+static esp_err_t http_resp_dir(httpd_req_t *req, const char *dirpath)
 {
     char entrypath[FILE_PATH_MAX];
     char entrysize[16];
@@ -184,24 +166,49 @@ esp_err_t state_post_handler(httpd_req_t *req) {
     uint8_t status_code;
     static robot_state prev_state = STOP;
 
-    if (sscanf(buf, "statusCode=%hhu", &status_code) == 1) {
-        cmd = (robot_state)status_code;
-        // Overflow if status_code is bigger than char
+    if (xSemaphoreTake(robot_state_mutex, pdMS_TO_TICKS(500))){
+        //TODO: Edit to use only one xSemaphoreGive
+        uint8_t parsed = sscanf(buf, "statusCode=%hhu&filename=%63s", &status_code, filename);
+        if (parsed >= 1) {
+            cmd = (robot_state)status_code;
+            // Overflow if status_code is bigger than char
+            
+            if (cmd == START) {
+                if (parsed != 2) {
+                    httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "Missing filename");
+                    xSemaphoreGive(robot_state_mutex);
+                    return ESP_FAIL;
+                }
+                
+                size_t len = strlen(filename);
+                if (len < 6 || strcmp(filename + len - 6, ".gcode") != 0) {
+                    httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "Invalid file type");
+                    xSemaphoreGive(robot_state_mutex);
+                    return ESP_FAIL;
+                }
 
-        if (cmd == START) {
-            ESP_LOGI(TAG, "START");
-        } else if (cmd == STOP) {
-            ESP_LOGI(TAG, "STOP");
-        } else if (cmd == PAUSE && prev_state == START){
-            ESP_LOGI(TAG, "PAUSE");
-        } else {
-            ESP_LOGW(TAG, "Unknown command");
+                current_state = START;
+                ESP_LOGI(TAG, "START file: %s", filename);
+            } else if (cmd == STOP) {
+                current_state = STOP;
+                ESP_LOGI(TAG, "STOP");
+            } else if (cmd == PAUSE && prev_state == START){
+                current_state = PAUSE;
+                ESP_LOGI(TAG, "PAUSE");
+            } else {
+                ESP_LOGW(TAG, "Unknown command");
+            }
+
+            prev_state = cmd;
+            httpd_resp_sendstr(req, "State updated successfully");
+            xSemaphoreGive(robot_state_mutex);
+            return ESP_OK;
         }
-
-        prev_state = cmd;
-        httpd_resp_sendstr(req, "State updated successfully");
-        return ESP_OK;
-    } 
+        xSemaphoreGive(robot_state_mutex);
+    } else {
+        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Mutex timeout");
+        return ESP_FAIL;
+    }
 
     httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "Invalid state data");
     return ESP_FAIL;
@@ -209,7 +216,7 @@ esp_err_t state_post_handler(httpd_req_t *req) {
 
 esp_err_t manual_control_post_handler(httpd_req_t *req) {
     /// Variables for angles: base, arm, tool, and coordinates: x, y
-    controls ctrl = {0};
+    mov_controls ctrl = {0};
     char buf[DATA_BUFSIZE];
     int size;
     uint8_t res;
@@ -237,6 +244,8 @@ esp_err_t manual_control_post_handler(httpd_req_t *req) {
     );
 
     if (res == 5){
+        xQueueSend(manual_mov_queue, &ctrl, pdMS_TO_TICKS(100));
+
         ESP_LOGI(TAG, "Parsed successfully: base=%d, arm=%d, tool=%d, x=%d, y=%d",
             ctrl.base_angle, 
             ctrl.arm_angle, 
@@ -253,24 +262,28 @@ esp_err_t manual_control_post_handler(httpd_req_t *req) {
     return ESP_FAIL;
 }
 
-esp_err_t accel_data_get_handler(httpd_req_t *req){
-    accel_data *data = (accel_data*)req->user_ctx;
+esp_err_t accel_vector_get_handler(httpd_req_t *req){
+    accel_vector *data = (accel_vector*)req->user_ctx;
     if (data == NULL) {
         httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Missing data");
         return ESP_FAIL;
     }
 
     char buf[DATA_BUFSIZE];
-    snprintf(buf, sizeof(buf), "x_accel=%hu&y_accel=%hu&z_accel=%hu",
-        data->x_accel, 
-        data->y_accel, 
-        data->z_accel
-    );
-   ESP_LOGI(TAG, "Sent successfully: x_accel=%d, y_accel=%d, z_accel=%d",
-            data->x_accel,
-            data->y_accel,
+
+    if(xSemaphoreTake(imu_readings_mutex, pdMS_TO_TICKS(200))){
+        snprintf(buf, sizeof(buf), "x_accel=%hd&y_accel=%hd&z_accel=%hd",
+            data->x_accel, 
+            data->y_accel, 
             data->z_accel
-    );
+        );
+        xSemaphoreGive(imu_readings_mutex);
+    } else {
+        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Mutex timeout");
+        return ESP_FAIL;
+    }
+
+    ESP_LOGI(TAG, "Sent successfully: %s", buf);
 
 #ifdef CONFIG_EXAMPLE_HTTPD_CONN_CLOSE_HEADER
     httpd_resp_set_hdr(req, "Connection", "close");
@@ -298,7 +311,7 @@ static esp_err_t download_get_handler(httpd_req_t *req)
 
     /* If name has trailing '/', respond with directory contents */
     if (filename[strlen(filename) - 1] == '/') {
-        return http_resp_dir_html(req, filepath);
+        return http_resp_dir(req, filepath);
     }
 
     if (stat(filepath, &file_stat) == -1) {
@@ -308,42 +321,47 @@ static esp_err_t download_get_handler(httpd_req_t *req)
         return ESP_FAIL;
     }
 
-    fd = fopen(filepath, "r");
-    if (!fd) {
-        ESP_LOGE(TAG, "Failed to read existing file : %s", filepath);
-        /* Respond with 500 Internal Server Error */
-        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to read existing file");
-        return ESP_FAIL;
-    }
-
-    ESP_LOGI(TAG, "Sending file : %s (%ld bytes)...", filename, file_stat.st_size);
-    set_content_type_from_file(req, filename);
-
-    /* Retrieve the pointer to scratch buffer for temporary storage */
-    char *chunk = ((struct file_server_data *)req->user_ctx)->scratch;
-    size_t chunksize;
-    do {
-        /* Read file in chunks into the scratch buffer */
-        chunksize = fread(chunk, 1, SCRATCH_BUFSIZE, fd);
-
-        if (chunksize > 0) {
-            /* Send the buffer contents as HTTP response chunk */
-            if (httpd_resp_send_chunk(req, chunk, chunksize) != ESP_OK) {
-                fclose(fd);
-                ESP_LOGE(TAG, "File sending failed!");
-                /* Abort sending file */
-                httpd_resp_sendstr_chunk(req, NULL);
-                /* Respond with 500 Internal Server Error */
-                httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to send file");
-               return ESP_FAIL;
-           }
+    if (xSemaphoreTake(sd_card_mutex, pdMS_TO_TICKS(portMAX_DELAY))){
+        fd = fopen(filepath, "r");
+        if (!fd) {
+            ESP_LOGE(TAG, "Failed to read existing file : %s", filepath);
+            /* Respond with 500 Internal Server Error */
+            httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to read existing file");
+            xSemaphoreGive(sd_card_mutex);
+            return ESP_FAIL;
         }
 
-        /* Keep looping till the whole file is sent */
-    } while (chunksize != 0);
+        ESP_LOGI(TAG, "Sending file : %s (%ld bytes)...", filename, file_stat.st_size);
+        set_content_type_from_file(req, filename);
 
-    /* Close file after sending complete */
-    fclose(fd);
+        /* Retrieve the pointer to scratch buffer for temporary storage */
+        char *chunk = ((struct file_server_data *)req->user_ctx)->scratch;
+        size_t chunksize;
+        do {
+            /* Read file in chunks into the scratch buffer */
+            chunksize = fread(chunk, 1, SCRATCH_BUFSIZE, fd);
+
+            if (chunksize > 0) {
+                /* Send the buffer contents as HTTP response chunk */
+                if (httpd_resp_send_chunk(req, chunk, chunksize) != ESP_OK) {
+                    fclose(fd);
+                    ESP_LOGE(TAG, "File sending failed!");
+                    /* Abort sending file */
+                    httpd_resp_sendstr_chunk(req, NULL);
+                    /* Respond with 500 Internal Server Error */
+                    httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to send file");
+                    xSemaphoreGive(sd_card_mutex);
+                return ESP_FAIL;
+                }
+            }
+
+            /* Keep looping till the whole file is sent */
+        } while (chunksize != 0);
+
+        /* Close file after sending complete */
+        fclose(fd);
+        xSemaphoreGive(sd_card_mutex);
+    }
     ESP_LOGI(TAG, "File sending complete");
 
     /* Respond with an empty chunk to signal HTTP response completion */
@@ -365,7 +383,6 @@ static esp_err_t upload_post_handler(httpd_req_t *req)
     /* Note sizeof() counts NULL termination hence the -1 */
     const char *filename = get_path_from_uri(filepath, ((struct file_server_data *)req->user_ctx)->base_path,
                                              req->uri + sizeof("/upload") - 1, sizeof(filepath));
-
     if (!filename) {
         /* Respond with 500 Internal Server Error */
         httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Filename too long");
@@ -398,65 +415,71 @@ static esp_err_t upload_post_handler(httpd_req_t *req)
         return ESP_FAIL;
     }
 
-    fd = fopen(filepath, "w");
-    if (!fd) {
-        ESP_LOGE(TAG, "Failed to create file : %s", filepath);
-        /* Respond with 500 Internal Server Error */
-        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to create file");
-        return ESP_FAIL;
-    }
+    if (xSemaphoreTake(sd_card_mutex, pdMS_TO_TICKS(portMAX_DELAY))){
+        fd = fopen(filepath, "w");
+        if (!fd) {
+            ESP_LOGE(TAG, "Failed to create file : %s", filepath);
+            /* Respond with 500 Internal Server Error */
+            httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to create file");
+            xSemaphoreGive(sd_card_mutex);
+            return ESP_FAIL;
+        }
 
-    ESP_LOGI(TAG, "Receiving file : %s...", filename);
+        ESP_LOGI(TAG, "Receiving file : %s...", filename);
 
-    /* Retrieve the pointer to scratch buffer for temporary storage */
-    char *buf = ((struct file_server_data *)req->user_ctx)->scratch;
-    int received;
+        /* Retrieve the pointer to scratch buffer for temporary storage */
+        char *buf = ((struct file_server_data *)req->user_ctx)->scratch;
+        int received;
 
-    /* Content length of the request gives
-     * the size of the file being uploaded */
-    int remaining = req->content_len;
+        /* Content length of the request gives
+        * the size of the file being uploaded */
+        int remaining = req->content_len;
 
-    while (remaining > 0) {
+        while (remaining > 0) {
 
-        ESP_LOGI(TAG, "Remaining size : %d", remaining);
-        /* Receive the file part by part into a buffer */
-        if ((received = httpd_req_recv(req, buf, MIN(remaining, SCRATCH_BUFSIZE))) <= 0) {
-            if (received == HTTPD_SOCK_ERR_TIMEOUT) {
-                /* Retry if timeout occurred */
-                continue;
+            ESP_LOGI(TAG, "Remaining size : %d", remaining);
+            /* Receive the file part by part into a buffer */
+            if ((received = httpd_req_recv(req, buf, MIN(remaining, SCRATCH_BUFSIZE))) <= 0) {
+                if (received == HTTPD_SOCK_ERR_TIMEOUT) {
+                    /* Retry if timeout occurred */
+                    continue;
+                }
+
+                /* In case of unrecoverable error,
+                * close and delete the unfinished file*/
+                fclose(fd);
+                unlink(filepath);
+                xSemaphoreGive(sd_card_mutex);
+
+                ESP_LOGE(TAG, "File reception failed!");
+                /* Respond with 500 Internal Server Error */
+                httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to receive file");
+                return ESP_FAIL;
             }
 
-            /* In case of unrecoverable error,
-             * close and delete the unfinished file*/
-            fclose(fd);
-            unlink(filepath);
+            /* Write buffer content to file on storage */
+            if (received && (received != fwrite(buf, 1, received, fd))) {
+                /* Couldn't write everything to file!
+                * Storage may be full? */
+                fclose(fd);
+                unlink(filepath);
+                xSemaphoreGive(sd_card_mutex);
 
-            ESP_LOGE(TAG, "File reception failed!");
-            /* Respond with 500 Internal Server Error */
-            httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to receive file");
-            return ESP_FAIL;
+                ESP_LOGE(TAG, "File write failed!");
+                /* Respond with 500 Internal Server Error */
+                httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to write file to storage");
+                return ESP_FAIL;
+            }
+
+            /* Keep track of remaining size of
+            * the file left to be uploaded */
+            remaining -= received;
         }
 
-        /* Write buffer content to file on storage */
-        if (received && (received != fwrite(buf, 1, received, fd))) {
-            /* Couldn't write everything to file!
-             * Storage may be full? */
-            fclose(fd);
-            unlink(filepath);
-
-            ESP_LOGE(TAG, "File write failed!");
-            /* Respond with 500 Internal Server Error */
-            httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Failed to write file to storage");
-            return ESP_FAIL;
-        }
-
-        /* Keep track of remaining size of
-         * the file left to be uploaded */
-        remaining -= received;
+        /* Close file upon upload completion */
+        fclose(fd);
+        xSemaphoreGive(sd_card_mutex);
     }
-
-    /* Close file upon upload completion */
-    fclose(fd);
     ESP_LOGI(TAG, "File reception complete");
 
     /* Redirect onto root to see the updated file list */
@@ -501,7 +524,10 @@ static esp_err_t delete_post_handler(httpd_req_t *req)
 
     ESP_LOGI(TAG, "Deleting file : %s", filename);
     /* Delete file */
-    unlink(filepath);
+    if(xSemaphoreTake(sd_card_mutex, pdMS_TO_TICKS(portMAX_DELAY))){
+        unlink(filepath);
+        xSemaphoreGive(sd_card_mutex);
+    }
 
     /* Redirect onto root to see the updated file list */
     httpd_resp_set_status(req, "303 See Other");
@@ -546,17 +572,12 @@ esp_err_t start_file_server(const char *base_path)
         return ESP_FAIL;
     }
 
-    static accel_data data;
-    data.x_accel = 9;
-    data.y_accel = 1;
-    data.z_accel = 1;
-
     /* URI handler for getting uploaded files */
     httpd_uri_t get_data = {
         .uri       = "/accel_data",
         .method    = HTTP_GET,
-        .handler   = accel_data_get_handler,
-        .user_ctx  = &data
+        .handler   = accel_vector_get_handler,
+        .user_ctx  = &imu_readings
     };
     httpd_register_uri_handler(server, &get_data);
 
@@ -576,13 +597,13 @@ esp_err_t start_file_server(const char *base_path)
     };
     httpd_register_uri_handler(server, &set_state);
 
-    httpd_uri_t set_controls = {
+    httpd_uri_t set_mov_controls = {
         .uri       = "/controls",
         .method    = HTTP_POST,
         .handler   = manual_control_post_handler,
         .user_ctx  = server_data
     };
-    httpd_register_uri_handler(server, &set_controls);
+    httpd_register_uri_handler(server, &set_mov_controls);
 
     /* URI handler for uploading files to server */
     httpd_uri_t file_upload = {
